pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grok-super-agent"
version = "0.2.0"
description = "Grok Super Agent — 멀티 AI 오케스트레이션 + Computer Use. x.ai/Tesla ecosystem."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Grok Super Agent" }]
dependencies = [
    "xai-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
    "computeruseprotocol[mcp]>=0.1.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
grok-super-agent = "src.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
